% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppendIDfunctions.R
\name{cy_append_ids}
\alias{cy_append_ids}
\title{Append Monadic Country-Year Identifiers}
\usage{
cy_append_ids(df, breaks = T)
}
\arguments{
\item{df}{A dataframe object with monadic country data which includes a column designating country names or identifiers for each observation and a column designating years for each observation.
Supplied country identifiers can be provided as a column containing country names, or Gleditch-Ward or Correlates of War abbreviations or identifiers. Function is not case-sensitive for country column names, as long as column name includes "country".}

\item{breaks}{TRUE or FALSE, if TRUE ID joining process will be interrupted at instances where GWNO codes cannot be identified for a given country-year observation. Returns a vector of missing GWNO codes.}
}
\value{
A dataframe object with monadic country data, updated to standardize current country names based on Gleditsch-Ward number and year observations (Gleditsch and War 1999).
Country names will be returned as the column "country" and years as "year", Previous identifiers returned as "countryname_raw".
Dataframe will also include the following additional country-year identifiers: Gleditsch-Ward Number [gwno], Gleditsch-Ward Abbreviation [gwabbrev], Correlates of War Country Code [ccode], International Finanical Statistics (IFS) number [ifscode], and IFS abbreviation [ifs].
}
\description{
Standardizes a dataframe with monadic country-year observations by creating a "country" column based on Gleditsch-Ward (GW) country-year observations.
A master dataframe of Gleditch-Ward country-year observations is used to standardize country-year identifiers across observations, allowing for merging of datasets from sources using different country name coding schemes.
In addition to GW country names and country codes (gwno), adds additional columns with other potential country-year identifiers including Correlates of War (COW) and International Finanical Statistics (IFS) codes.
Supplied dataframe must include columns designating each observation's country identifier and the year of observation.
Dataframe must include single observations for each country-year without duplicated observations.
}
\examples{
data_cy <- cy_append_ids(data, breaks=T)
}
