% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppendIDfunctions.R
\name{dyad_append_ids}
\alias{dyad_append_ids}
\title{Append Dyadic Country-Year Identifiers}
\usage{
dyad_append_ids(df, breaks = T)
}
\arguments{
\item{df}{A dataframe object with dyadic country data which includes columns designating country names or identifiers for each country in the dyad and a column designating years for each observation.
Supplied country identifiers can be provided as a column containing country names, or Gleditch-Ward or Correlates of War abbreviations or identifiers.
Dyadic country identifiers can be differentiated numerically ("Country1" and "Country2") or alphabetically ("CountryA" and "CountryB").
Function is not case-sensitive for country column names, as long as column name includes "country".}

\item{breaks}{TRUE or FALSE, if TRUE ID joining process will be interrupted at instances where GWNO codes cannot be identified for a given country-year observation. Returns a vector of missing GWNO codes.}
}
\value{
A dataframe object with dyadic country data, updated to standardize current country names based on Gleditsch-Ward number and year observations (Gleditsch and War 1999).
Country names will be returned as the columns "repcountry" and "parcountry" and years as "year", Previous identifiers returned as "repcountryname_raw" and "parcountryname_raw".
Dataframe will also include the following additional country-year identifiers: Gleditsch-Ward Number [gwno], Gleditsch-Ward Abbreviation [gwabbrev], Correlates of War Country Code [ccode], International Finanical Statistics (IFS) number [ifscode], and IFS abbreviation [ifs].
}
\description{
Standardizes a dataframe with dyadic country-year observations by creating "repcountry" and "parcountry" columns based on Gleditsch-Ward (GW) country-year observations.
A master dataframe of Gleditch-Ward country-year observations is used to standardize country-year identifiers across observations, allowing for merging of datasets from sources using different country name coding schemes.
In addition to GW country names and country codes (gwno), adds additional columns with other potential country-year identifiers including Correlates of War (COW) and International Finanical Statistics (IFS) codes.
Supplied dataframe must include columns designating observations for each dyad member's country identifier and the year of observation.
Dataframe must include single observations for each repcountry-parcountry-year without duplicated observations.
}
\examples{
data_dyad <- dyad_append_ids(data, breaks = F)
}
